requirements/dev.txt
# API Framework
fastapi>=0.100.0
uvicorn[standard]>=0.22.0
pydantic[email]>=2.0.0
pydantic-settings>=2.0.0
python-multipart  # for OAuth2 form handling
python-jose[cryptography]  # for JWT
passlib[bcrypt]  # for password hashing
redis>=5.0.0  # for rate limiting and caching

# Database
SQLAlchemy>=2.0.0
alembic>=1.11.0
psycopg2-binary  # for PostgreSQL
mysqlclient  # for MySQL
pymysql  # for MySQL

# Testing
pytest>=7.3.1
pytest-cov>=4.1.0
httpx>=0.24.1  # for TestClient
factory-boy>=3.3.0  # for test data generation

# Linting and formatting
black>=23.3.0
flake8>=6.0.0
mypy>=1.3.0
isort>=5.12.0

# Development tools
ipython  # for better REPL
debugpy  # for debugging